# docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /todos:
    post:
      description: Creates a new todo item for the authenticated user
      operationId: TodosController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTodoDto'
      responses:
        '201':
          description: Todo created successfully
        '400':
          description: Invalid input data
        '401':
          description: Unauthorized
      security:
        - bearer: []
      summary: Create a new todo
      tags:
        - todos
    get:
      description: >-
        Retrieves all todos. Regular users see only their own todos, while
        admins and sysadmins see all todos.
      operationId: TodosController_findAll
      parameters: []
      responses:
        '200':
          description: List of todos retrieved successfully
        '401':
          description: Unauthorized
      security:
        - bearer: []
      summary: Get all todos
      tags:
        - todos
  /todos/{id}:
    get:
      description: >-
        Retrieves a specific todo by its ID. Regular users can only access their
        own todos.
      operationId: TodosController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Todo retrieved successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Not authorized to access this todo
        '404':
          description: Todo not found
      security:
        - bearer: []
      summary: Get a todo by ID
      tags:
        - todos
    patch:
      description: Updates a todo by its ID. Regular users can only update their own todos.
      operationId: TodosController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTodoDto'
      responses:
        '200':
          description: Todo updated successfully
        '400':
          description: Invalid input data
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Not authorized to update this todo
        '404':
          description: Todo not found
      security:
        - bearer: []
      summary: Update a todo
      tags:
        - todos
    delete:
      description: Deletes a todo by its ID. Regular users can only delete their own todos.
      operationId: TodosController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: Todo deleted successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Not authorized to delete this todo
        '404':
          description: Todo not found
      security:
        - bearer: []
      summary: Delete a todo
      tags:
        - todos
  /admin/users:
    get:
      description: Retrieves all users in the system. Requires admin or sysadmin role.
      operationId: AdminController_getAllUsers
      parameters: []
      responses:
        '200':
          description: List of users retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponseDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Requires admin or sysadmin role
      security:
        - bearer: []
      summary: Get all users
      tags:
        - admin
  /admin/users/{id}:
    get:
      description: Retrieves a specific user by their ID. Requires admin or sysadmin role.
      operationId: AdminController_getUserById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: User retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Requires admin or sysadmin role
        '404':
          description: User not found
      security:
        - bearer: []
      summary: Get user by ID
      tags:
        - admin
    patch:
      description: >-
        Updates a user by their ID. Can update email, password, full name, role,
        and email verification status. Requires sysadmin role.
      operationId: AdminController_updateUser
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
        '400':
          description: Invalid input data
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Requires sysadmin role
        '404':
          description: User not found
      security:
        - bearer: []
      summary: Update user
      tags:
        - admin
    delete:
      description: Deletes a user by their ID. Requires sysadmin role.
      operationId: AdminController_deleteUser
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: User deleted successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Requires sysadmin role
        '404':
          description: User not found
      security:
        - bearer: []
      summary: Delete user
      tags:
        - admin
...
info:
  title: Todo App API
  description: >-
    REST API for Todo Application with authentication and role-based access
    control
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    CreateTodoDto:
      type: object
      properties:
        description:
          type: string
          description: Description of the todo item
          example: Buy groceries for the week
        dueDate:
          format: date-time
          type: string
          description: Due date for the todo item
          example: '2025-12-31T23:59:59.000Z'
        priority:
          type: string
          description: Priority level of the todo item
          enum:
            - low
            - medium
            - high
          example: medium
      required:
        - description
    UpdateTodoDto:
      type: object
      properties:
        description:
          type: string
          description: Description of the todo item
          example: Buy groceries for the week
        dueDate:
          format: date-time
          type: string
          description: Due date for the todo item
          example: '2025-12-31T23:59:59.000Z'
        priority:
          type: string
          description: Priority level of the todo item
          enum:
            - low
            - medium
            - high
          example: medium
    UserResponseDto:
      type: object
      properties:
        id:
          type: string
          description: User ID
          example: uuid-123
        email:
          type: string
          description: User email address (primary)
          example: user@example.com
        fullName:
          type: string
          description: User full name
          example: John Doe
        role:
          type: string
          description: User role
          example: guest
        emailVerifiedAt:
          type: object
          description: Email verification timestamp
          example: '2025-01-01T00:00:00.000Z'
          nullable: true
        localUsername:
          type: object
          description: Local username (could be email or username)
          example: john.doe
          nullable: true
        googleEmail:
          type: object
          description: Google account email
          example: user@gmail.com
          nullable: true
        msEmail:
          type: object
          description: Microsoft account email
          example: user@outlook.com
          nullable: true
        createdAt:
          format: date-time
          type: string
          description: User creation timestamp
          example: '2025-01-01T00:00:00.000Z'
        updatedAt:
          format: date-time
          type: string
          description: User last update timestamp
          example: '2025-01-01T00:00:00.000Z'
      required:
        - id
        - email
        - fullName
        - role
        - emailVerifiedAt
        - localUsername
        - googleEmail
        - msEmail
        - createdAt
        - updatedAt
    UpdateUserDto:
      type: object
      properties:
        email:
          type: string
          description: User email address
          example: user@example.com
        password:
          type: string
          description: User password (minimum 8 characters)
          example: newPassword123
          minLength: 8
          maxLength: 128
        fullName:
          type: string
          description: User full name
          example: John Doe
          minLength: 1
          maxLength: 255
        role:
          type: string
          description: User role
          enum:
            - guest
            - admin
            - sysadmin
          example: guest
        emailVerifiedAt:
          type: object
          description: Email verification timestamp (ISO 8601 format, or null to unverify)
          example: '2025-01-01T00:00:00.000Z'
...